feat: run Fynd's own command line with algorithms a caller brought
WHY:
`serve` lived in `main.rs`, so a binary that wanted Fynd's command line with
its own algorithms registered had to reimplement it — roughly 390 lines of
argument handling, tracing and metrics setup that would then drift from the
original every time a flag changed.

HOW:
- Move `setup_solver` and `run_solver` out of `main.rs` into `fynd::serve`,
  taking an `AlgorithmRegistry` and handing it to the builder.
- Split `serve` out of `run_solver`: `run_solver` builds the runtime and
  installs the global tracing subscriber and metrics recorder, `serve` takes
  over none of them, so a caller that already owns those can still embed.
  Say which is which in both doc comments.
- Expose `fynd::cli`, so a downstream binary parses the same `Cli`, and show
  the whole path in the crate doc.
- Keep `resolve_tycho_url` and `resolve_rpc_url` `pub(crate)`: their only
  caller is `commands::derive_connector_tokens`, in this crate.

ADDITIONAL NOTES:
`main.rs` keeps only `fn main` and passes an empty registry, so the shipped
binary behaves exactly as before. The diff is large but almost entirely a
move; the registry parameter and the runtime split are the only edits inside
what moved.
