+++ src/lib.rs
@@ -14,6 +14,24 @@
 //! use fynd::rpc::builder::FyndRPCBuilder;
 //! use fynd::core::algorithm::Algorithm;
 //! ```
+//!
+//! To run Fynd's own command line with an algorithm of your own, parse [`cli::Cli`] and hand
+//! [`serve::run_solver`] a registry — the binary keeps every flag `fynd serve` has:
+//!
+//! ```rust,ignore
+//! let algorithms = fynd::core::AlgorithmRegistry::new().with_algorithm("mine", Mine::new)?;
+//! match fynd::cli::Cli::parse().command {
+//!     fynd::cli::Commands::Serve(args) => fynd::serve::run_solver(*args, algorithms)?,
+//!     _ => {}
+//! }
+//! ```
 
 pub use fynd_core as core;
 pub use fynd_rpc as rpc;
+
+/// Command-line arguments, so a binary embedding Fynd can parse the same ones.
+pub mod cli;
+/// Subcommands other than `serve`.
+pub mod commands;
+/// Running the solver, as `fynd serve` does.
+pub mod serve;
